move game of life files from thundermoon to game of life app
